move code from vectorize to multi
