+++ pkg/sql/plan2/function/builtin/multi/concat_ws_test.go
@@ -15,12 +15,39 @@
 package multi
 
 import (
+	"github.com/matrixorigin/matrixone/pkg/container/nulls"
 	"github.com/matrixorigin/matrixone/pkg/container/types"
 	"github.com/matrixorigin/matrixone/pkg/container/vector"
 	"github.com/matrixorigin/matrixone/pkg/sql/testutil"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
+
+func TestConcatWsInputBytes(t *testing.T) {
+	inputBytes := make([]*types.Bytes, 2)
+	inputNsps := make([]*nulls.Nulls, 2)
+	inputBytes[0] = &types.Bytes{Data: []byte("hellogutenguten"),
+		Lengths: []uint32{5, 0, 5, 5},
+		Offsets: []uint32{0, 5, 5, 10},
+	}
+	inputBytes[1] = &types.Bytes{Data: []byte("hellogutenguten"),
+		Lengths: []uint32{5, 0, 5, 5},
+		Offsets: []uint32{0, 5, 5, 10},
+	}
+	inputNsps[0] = new(nulls.Nulls)
+	inputNsps[1] = new(nulls.Nulls)
+	nulls.Add(inputNsps[0], uint64(1))
+	nulls.Add(inputNsps[1], uint64(1))
+	resultValues := types.Bytes{
+		Data:    make([]byte, 0),
+		Lengths: make([]uint32, 4),
+		Offsets: make([]uint32, 4),
+	}
+	concatWsInputBytes(inputBytes, inputNsps, []byte("---"), 4, &resultValues)
+	require.Equal(t, []byte("hello---helloguten---gutenguten---guten"), resultValues.Data)
+	require.Equal(t, []uint32{0, 13, 13, 26}, resultValues.Offsets)
+	require.Equal(t, []uint32{13, 0, 13, 13}, resultValues.Lengths)
+}
 
 func TestConcat_ws(t *testing.T) {
 	vector0 := testutil.MakeScalarVarchar("---", 4)
